Accept version number in configuration
Before this change, we would only accept a pack format when creating a new project with JMC, which was not very convenient if you had not memorized the relevant format. 

With this change, we will try to break down the input in a tuple to find the correct pack format for a version.