Fix test cases (#52)
* Fix `RightClick.Setup()`

`@s __item_id__` is never set by anything, so using as as the input for switch is a mistake.

* Fix test cases for pre-macro switch

The changes in #50 affected not just the test case for `switch` itself, but also for all built-in functions that use switch.

TODO: make some way to change pack version in test environment, so we can also cover 1.20.2+ switch